+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### unreleased
+#### Added
+- option to apply custom screen color
+
+---
+
 ### 2.2.1 (2025-04-24)
 #### Fixed
 - more cleanup of events specifier